bpf: add BPF_PROG_RUN_TEST tests
This patch introduces bpf testing facilities using BPF_PROG_RUN_TEST. We
use the cilium/ebpf library to load and test the programs.

cilium/ebpf does not seem to support  the .pinning field for maps when
they are read from the corresponding ELF section. Not sure how easy it
would be to add such support, but get something working without changing
ebpf we patch the ELF file and zero the unsupported fields out.

There are three tests currently. The first two are just sanity checks
that everything works as expected. The last one prints the timeout of a
CT entry when a SYN TCP packet is sent and a TCP RST packet is received.
It is intentended for a subsequent patch.

The tests can be run as follows from the `bpf/test/` directory:
```
$ make -C ../ tests/bpf_ct_tests.o
$ sudo go run bpf-prog-test-run/bpf-prog-test-run.go

Log Messages:
 MARK 0x0 FROM 0 DEBUG: Conntrack lookup 1/2: src=10.3.0.10:3010 dst=10.3.0.20:3020
 MARK 0x0 FROM 0 DEBUG: Conntrack lookup 2/2: nexthdr=6 flags=1
 MARK 0x0 FROM 0 DEBUG: CT verdict: New, revnat=0
 MARK 0x0 FROM 0 DEBUG: Conntrack create: proxy-port=0 revnat=0 src-identity=0 lb=0.0.0.0
CT Entries:
 key=10.3.0.20:49675 --> 10.3.0.10:52235, 6, 0 val=expires=1262314
 RxPackets=0 RxBytes=0 RxFlagsSeen=0x00 LastRxReport=0 TxPackets=1
 TxBytes=64 TxFlagsSeen=0x02 LastTxReport=1262254 Flags=0x0000 [ ]
 RevNAT=0 SourceSecurityID=0 IfIndex=0
Log Messages:
 MARK 0x0 FROM 0 DEBUG: Conntrack lookup 1/2: src=10.3.0.20:3020 dst=10.3.0.10:3010
 MARK 0x0 FROM 0 DEBUG: Conntrack lookup 2/2: nexthdr=6 flags=0
 MARK 0x0 FROM 0 DEBUG: CT entry found lifetime=1262314, revnat=0
 MARK 0x0 FROM 0 DEBUG: CT verdict: Reply, revnat=0
 MARK 0x0 FROM 0 DEBUG: Unknown message type=0 arg1=1000 arg2=2
CT Entries:
 key=10.3.0.20:3010 --> 10.3.0.10:3020, 6, 0 val=expires=1283854
 RxPackets=0 RxBytes=0 RxFlagsSeen=0x04 LastRxReport=1262254 TxPackets=1
 TxBytes=64 TxFlagsSeen=0x02 LastTxReport=1262254 Flags=0x0011 [
 RxClosing SeenNonSyn ] RevNAT=0 SourceSecurityID=0 IfIndex=0
Entry expires in 21600s
```

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>